dashboard/requirements.txt
#https://pypi.org/project/google-api-python-client/
google-api-python-client==2.51.0

appengine-python-standard>=0.2.2
click==7.1.2
Flask==1.1.2
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-logging==3.1.2
google-python-cloud-debugger==3.2
gunicorn==20.1.0

#https://pypi.org/project/httplib2/
httplib2==0.20.4

ijson==3.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1

protobuf~=3.19.0
PyCryptodome==3.14.1

#https://pypi.org/project/pyparsing/
pyparsing==3.0.9

six==1.15.0

uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==1.0.1
